# Conflicts:
#	modules/swagger-codegen/src/main/resources/aspnetcore/project.json.mustache

+++ samples/server/petstore/rust-server/src/models.rs
@@ -322,7 +322,7 @@ impl EnumArrays {
 /// which helps with FFI.
 #[allow(non_camel_case_types)]
 #[repr(C)]
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize, Eq, Ord)]
 pub enum EnumClass { 
     #[serde(rename = "_abc")]
     _ABC,
@@ -392,11 +392,11 @@ impl EnumTest {
 pub struct FormatTest {
     #[serde(rename = "integer")]
     #[serde(skip_serializing_if="Option::is_none")]
-    pub integer: Option<i32>,
+    pub integer: Option<u8>,
 
     #[serde(rename = "int32")]
     #[serde(skip_serializing_if="Option::is_none")]
-    pub int32: Option<i32>,
+    pub int32: Option<u32>,
 
     #[serde(rename = "int64")]
     #[serde(skip_serializing_if="Option::is_none")]
@@ -603,7 +603,7 @@ pub struct Name {
 
     #[serde(rename = "123Number")]
     #[serde(skip_serializing_if="Option::is_none")]
-    pub _123_number: Option<i32>,
+    pub _123_number: Option<isize>,
 
 }
 
@@ -739,7 +739,7 @@ impl OuterComposite {
 /// which helps with FFI.
 #[allow(non_camel_case_types)]
 #[repr(C)]
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize, Eq, Ord)]
 pub enum OuterEnum { 
     #[serde(rename = "placed")]
     PLACED,
